requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
environs==9.5.0
Flask==2.3.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
packaging==23.1
psycopg2==2.9.6
pydantic==1.10.9
PyJWT==2.7.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==1.4.28
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
